requirements.txt
astroid==2.2.5
blinker==1.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Moment==0.7.0
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.3.1
PyMySQL==0.9.3
six==1.12.0
soupsieve==1.9.1
SQLAlchemy==1.3.4
typed-ast==1.4.0
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.4
wrapt==1.11.1
WTForm==1.0
WTForms==2.2.1
PyJWT==1.7.1